Break down method-invocation files into several files
