+++ src/parser/method/parameter.ts
@@ -1,0 +1,39 @@
+import * as T from '../../types';
+import { TokenTracker, Scope } from '../utils';
+
+export function parseParameters(
+  tt: TokenTracker,
+  parseExpr: T.ExpressionParsingFunction,
+  scope: Scope,
+): Array<T.Expr> {
+  const params: Array<T.Expr> = [];
+
+  while (true) {
+    let parameterExpr: T.Expr;
+    parameterExpr = parseParameter(tt, parseExpr, scope);
+    params.push(parameterExpr);
+
+    if (tt.isOneOf('rparen', 'newline')) break;
+    tt.next();
+  }
+
+  return params;
+}
+
+function parseParameter(
+  tt: TokenTracker,
+  parseExpr: T.ExpressionParsingFunction,
+  scope: Scope,
+): T.Expr {
+  const parameterExpr: T.AST = [];
+  let expr: T.Expr | undefined;
+
+  while (tt.isNotOneOf('newline', 'comma', 'rparen')) {
+    expr = parseExpr(undefined, { scope, ast: parameterExpr });
+    parameterExpr.push(expr);
+    tt.next();
+  }
+
+  let [result] = parameterExpr;
+  return result;
+}
